requirements.txt
pymupdf
transformers
qwen-vl-utils
pydantic
torch
torchvision
pillow
